Add forms, tables and vue vms for workshops. Add utility to convert 24 time into 12 with AM and PM. Handle delete requests for courses in case of success and error.

+++ app/Http/Controllers/CoursesController.php
@@ -31,7 +31,11 @@ class CoursesController extends Controller
 
     public function destroy(Course $course)
     {
-      $course->delete();
-      return ['message' => 'Curso Eliminado!', 'object' => $course];
+      if(!$course->hasWorkshops()) {
+        $course->delete();
+        return ['message' => 'Curso Eliminado!', 'object' => $course];
+      } else {
+        return response()->json(['message' => 'No se pudo eliminar el curso!'], 422);
+      }
     }
 }
